# Conflicts:
#	com.mire.cinema/src/main/resources/application.properties
#	com.mire.cinema/src/main/webapp/WEB-INF/header.jsp
#	com.mire.cinema/src/main/webapp/member/memberInfo.jsp

+++ com.mire.cinema/src/main/java/com/mire/cinema/service/MemberServiceImpl.java
@@ -7,15 +7,14 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.mire.cinema.domain.member.Member;
-import com.mire.cinema.domain.member.MemberLoginDTO;
-import com.mire.cinema.domain.member.MemberUpdateDTO;
+import com.mire.cinema.domain.member.MemberDTO;
 import com.mire.cinema.domain.member.TokenDTO;
+import com.mire.cinema.exception.ErrorMsg;
 import com.mire.cinema.repository.MemberMapper;
 import com.mire.cinema.security.JwtTokenProvider;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-
 
 @Slf4j
 @Service
@@ -36,49 +35,56 @@ public class MemberServiceImpl implements MemberService {
 	@Override
 	public Member findMember(String memberId) {
 
-		return memberMapper.selectMember(memberId);
+		Member member = memberMapper.selectMember(memberId);
+
+		return member;
 	}
 
 	@Override
-	public void modifyMember(MemberUpdateDTO dto) {
+	public void modifyMember(MemberDTO.Update dto) {
 		Member member = memberMapper.selectMember(dto.getMemberId());
+
 		if (member == null || !member.getMemberPasswd().equals(dto.getMemberPasswd())) {
-			new IllegalArgumentException("아이디와 비밀번호가 일치하지 않습니다.");
-			return;
+			throw new IllegalArgumentException(ErrorMsg.USERINFO);
 
 		}
 
-		if (dto.getMemberEmail() == null) {
-			dto.setMemberEmail(member.getMemberEmail());
-		}
-		if (dto.getMemberPhone() == null) {
-			dto.setMemberPhone(member.getMemberPhone());
-		}
-
-		memberMapper.updateMember(member);
+		memberMapper.updateMember(dto);
 
 	}
 
 	@Override
 	public void removeMember(String memberId) {
 
+		Member member = findMember(memberId);
+
+		if (member == null) {
+			System.out.println("예외호출");
+			throw new IllegalArgumentException(ErrorMsg.USERINFO);
+		}
+
 		memberMapper.deleteMember(memberId);
 
 	}
 
 	@Override
-	public TokenDTO loginMember(MemberLoginDTO dto) {
-		System.out.println(dto);
-		UsernamePasswordAuthenticationToken authenticationToken = new UsernamePasswordAuthenticationToken(dto.getMemberId(), dto.getMemberPasswd());
-		
+	public TokenDTO loginMember(MemberDTO.Login dto) {
+		Member member = memberMapper.selectMember(dto.getMemberId());
+
+		if (member == null || !member.getMemberPasswd().equals(dto.getMemberPasswd())) {
+			throw new IllegalArgumentException(ErrorMsg.USERINFO);
+
+		}
+
+		UsernamePasswordAuthenticationToken authenticationToken = new UsernamePasswordAuthenticationToken(
+				dto.getMemberId(), dto.getMemberPasswd());
+
 		Authentication authentication = authenticationManagerBuilder.getObject().authenticate(authenticationToken);
-		
+
 		TokenDTO tokenDTO = jwtTokenProvider.createToken(authentication);
-	
-		
-		
+
 		return tokenDTO;
-		
+
 	}
 
 }
